requirements.txt
fastapi==0.112.2               # Web API framework
uvicorn==0.30.6                # ASGI server for FastAPI

SQLAlchemy==2.0.32             # Database ORM
alembic==1.13.2                # Database migrations
aiosqlite==0.20.0              # Async SQLite connection

python-jose==3.3.0             # JWT tokens
bcrypt==4.0.1                  # Password hashing
passlib==1.7.4                 # Password tools
cryptography==43.0.0           # Encryption tools

httpx==0.27.2                  # Async HTTP requests
pytest==8.3.2                  # Testing framework
pytest-asyncio==0.24.0         # Async testing

jinja2==3.1.4                  # HTML templating
python-multipart==0.0.9        # Form and file uploads
aiofiles==24.1.0               # Async file operations

python-dotenv==1.0.1           # Load .env variables
markdown==3.7                  # Markdown to HTML
beautifulsoup4==4.12.3         # HTML parsing

google-generativeai==0.7.2     # Gemini API
langchain==0.2.15              # LLM workflows
langchain-google-genai==1.0.10 # Langchain Gemini integration
fastapi
uvicorn
sqlalchemy
python-dotenv

